compiles again..
[git-p4: depot-paths = "//depot/qt/main/": change = 35020]

+++ src/widgets/qmenubar_mac.cpp
@@ -187,7 +187,7 @@ bool QMenuBar::activate(MenuRef menu, short idx)
 	return FALSE;
     }
 
-    if(MacPopupBinding *mpb = pdict->find((int)menu)) {
+    if(MacPopupBinding *mpb = pdict->find((int)((short)menu))) {
 	MenuCommand cmd;
 	GetMenuItemCommandID(mpb->macpopup, idx, &cmd);
 	mpb->qpopup->activateItemAt(mpb->qpopup->indexOf(cmd));
